+++ src/db/redis.client.ts
@@ -1,7 +1,8 @@
 import { createClient, RedisClientType } from '@redis/client';
-import { Logger } from '../config/logger.config';
 import { BufferJSON } from '@whiskeysockets/baileys';
+
 import { Redis } from '../config/env.config';
+import { Logger } from '../config/logger.config';
 
 export class RedisCache {
   constructor() {
@@ -59,11 +60,7 @@ export class RedisCache {
       this.logger.verbose('writeData: ' + field);
       const json = JSON.stringify(data, BufferJSON.replacer);
 
-      return await this.client.hSet(
-        this.redisEnv.PREFIX_KEY + ':' + this.instanceName,
-        field,
-        json,
-      );
+      return await this.client.hSet(this.redisEnv.PREFIX_KEY + ':' + this.instanceName, field, json);
     } catch (error) {
       this.logger.error(error);
     }
@@ -72,10 +69,7 @@ export class RedisCache {
   public async readData(field: string) {
     try {
       this.logger.verbose('readData: ' + field);
-      const data = await this.client.hGet(
-        this.redisEnv.PREFIX_KEY + ':' + this.instanceName,
-        field,
-      );
+      const data = await this.client.hGet(this.redisEnv.PREFIX_KEY + ':' + this.instanceName, field);
 
       if (data) {
         this.logger.verbose('readData: ' + field + ' success');
@@ -92,10 +86,7 @@ export class RedisCache {
   public async removeData(field: string) {
     try {
       this.logger.verbose('removeData: ' + field);
-      return await this.client.hDel(
-        this.redisEnv.PREFIX_KEY + ':' + this.instanceName,
-        field,
-      );
+      return await this.client.hDel(this.redisEnv.PREFIX_KEY + ':' + this.instanceName, field);
     } catch (error) {
       this.logger.error(error);
     }
@@ -104,9 +95,7 @@ export class RedisCache {
   public async delAll(hash?: string) {
     try {
       this.logger.verbose('instance delAll: ' + hash);
-      const result = await this.client.del(
-        hash || this.redisEnv.PREFIX_KEY + ':' + this.instanceName,
-      );
+      const result = await this.client.del(hash || this.redisEnv.PREFIX_KEY + ':' + this.instanceName);
 
       return result;
     } catch (error) {