+++ src/whatsapp/abstract/abstract.router.ts
@@ -1,11 +1,13 @@
-import { InstanceDto } from '../dto/instance.dto';
-import { JSONSchema7 } from 'json-schema';
-import { Request } from 'express';
-import { validate } from 'jsonschema';
-import { BadRequestException } from '../../exceptions';
 import 'express-async-errors';
+
+import { Request } from 'express';
+import { JSONSchema7 } from 'json-schema';
+import { validate } from 'jsonschema';
+
 import { Logger } from '../../config/logger.config';
+import { BadRequestException } from '../../exceptions';
 import { GetParticipant, GroupInvite, GroupJid } from '../dto/group.dto';
+import { InstanceDto } from '../dto/instance.dto';
 
 type DataValidate<T> = {
   request: Request;
@@ -98,6 +100,7 @@ export abstract class RouterBroker {
 
   public async groupValidate<T>(args: DataValidate<T>) {
     const { request, ClassRef, schema, execute } = args;
+
     const instance = request.params as unknown as InstanceDto;
     const body = request.body;
 
@@ -188,9 +191,7 @@ export abstract class RouterBroker {
     const getParticipants = request.query as unknown as GetParticipant;
 
     if (!getParticipants?.getParticipants) {
-      throw new BadRequestException(
-        'The getParticipants needs to be informed in the query',
-      );
+      throw new BadRequestException('The getParticipants needs to be informed in the query');
     }
 
     const instance = request.params as unknown as InstanceDto;