+++ src/whatsapp/dto/chat.dto.ts
@@ -1,16 +1,7 @@
-import {
-  WAPrivacyOnlineValue,
-  WAPrivacyValue,
-  WAReadReceiptsValue,
-  proto,
-} from '@whiskeysockets/baileys';
+import { proto, WAPrivacyOnlineValue, WAPrivacyValue, WAReadReceiptsValue } from '@whiskeysockets/baileys';
 
 export class OnWhatsAppDto {
-  constructor(
-    public readonly jid: string,
-    public readonly exists: boolean,
-    public readonly name?: string,
-  ) {}
+  constructor(public readonly jid: string, public readonly exists: boolean, public readonly name?: string) {}
 }
 
 export class getBase64FromMediaMessageDto {