+++ src/whatsapp/services/chatwoot.service.ts
@@ -1,19 +1,17 @@
-import { InstanceDto } from '../dto/instance.dto';
-import path from 'path';
-import { ChatwootDto } from '../dto/chatwoot.dto';
-import { WAMonitoringService } from './monitor.service';
-import { Logger } from '../../config/logger.config';
 import ChatwootClient from '@figuro/chatwoot-sdk';
-import { createReadStream, readFileSync, unlinkSync, writeFileSync } from 'fs';
 import axios from 'axios';
 import FormData from 'form-data';
-import { SendTextDto } from '../dto/sendMessage.dto';
+import { createReadStream, readFileSync, unlinkSync, writeFileSync } from 'fs';
 import mimeTypes from 'mime-types';
-import { SendAudioDto } from '../dto/sendMessage.dto';
-import { SendMediaDto } from '../dto/sendMessage.dto';
-import { ROOT_DIR } from '../../config/path.config';
+import path from 'path';
+
 import { ConfigService, HttpServer } from '../../config/env.config';
-import { type } from 'os';
+import { Logger } from '../../config/logger.config';
+import { ROOT_DIR } from '../../config/path.config';
+import { ChatwootDto } from '../dto/chatwoot.dto';
+import { InstanceDto } from '../dto/instance.dto';
+import { SendAudioDto, SendMediaDto, SendTextDto } from '../dto/sendMessage.dto';
+import { WAMonitoringService } from './monitor.service';
 
 export class ChatwootService {
   private messageCacheFile: string;
@@ -23,10 +21,7 @@ export class ChatwootService {
 
   private provider: any;
 
-  constructor(
-    private readonly waMonitor: WAMonitoringService,
-    private readonly configService: ConfigService,
-  ) {
+  constructor(private readonly waMonitor: WAMonitoringService, private readonly configService: ConfigService) {
     this.messageCache = new Set();
   }
 
@@ -57,9 +52,7 @@ export class ChatwootService {
   private async getProvider(instance: InstanceDto) {
     this.logger.verbose('get provider to instance: ' + instance.instanceName);
     try {
-      const provider = await this.waMonitor.waInstances[
-        instance.instanceName
-      ].findChatwoot();
+      const provider = await this.waMonitor.waInstances[instance.instanceName].findChatwoot();
 
       if (!provider) {
         this.logger.warn('provider not found');
@@ -172,9 +165,7 @@ export class ChatwootService {
     });
 
     this.logger.verbose('check duplicate inbox');
-    const checkDuplicate = findInbox.payload
-      .map((inbox) => inbox.name)
-      .includes(inboxName);
+    const checkDuplicate = findInbox.payload.map((inbox) => inbox.name).includes(inboxName);
 
     let inboxId: number;
 
@@ -214,13 +205,7 @@ export class ChatwootService {
     this.logger.verbose('find contact in chatwoot and create if not exists');
     const contact =
       (await this.findContact(instance, '123456')) ||
-      ((await this.createContact(
-        instance,
-        '123456',
-        inboxId,
-        false,
-        'EvolutionAPI',
-      )) as any);
+      ((await this.createContact(instance, '123456', inboxId, false, 'EvolutionAPI')) as any);
 
     if (!contact) {
       this.logger.warn('contact not found');
@@ -425,22 +410,17 @@ export class ChatwootService {
 
       if (isGroup) {
         this.logger.verbose('get group name');
-        const group = await this.waMonitor.waInstances[
-          instance.instanceName
-        ].client.groupMetadata(chatId);
+        const group = await this.waMonitor.waInstances[instance.instanceName].client.groupMetadata(chatId);
 
         nameContact = `${group.subject} (GROUP)`;
 
         this.logger.verbose('find or create participant in chatwoot');
 
-        const picture_url = await this.waMonitor.waInstances[
-          instance.instanceName
-        ].profilePicture(body.key.participant.split('@')[0]);
-
-        const findParticipant = await this.findContact(
-          instance,
+        const picture_url = await this.waMonitor.waInstances[instance.instanceName].profilePicture(
           body.key.participant.split('@')[0],
         );
+
+        const findParticipant = await this.findContact(instance, body.key.participant.split('@')[0]);
 
         if (findParticipant) {
           if (!findParticipant.name || findParticipant.name === chatId) {
@@ -463,9 +443,7 @@ export class ChatwootService {
 
       this.logger.verbose('find or create contact in chatwoot');
 
-      const picture_url = await this.waMonitor.waInstances[
-        instance.instanceName
-      ].profilePicture(chatId);
+      const picture_url = await this.waMonitor.waInstances[instance.instanceName].profilePicture(chatId);
 
       const findContact = await this.findContact(instance, chatId);
 
@@ -510,8 +488,7 @@ export class ChatwootService {
         return null;
       }
 
-      const contactId =
-        contact?.payload?.id || contact?.payload?.contact?.id || contact?.id;
+      const contactId = contact?.payload?.id || contact?.payload?.contact?.id || contact?.id;
 
       if (!body.key.fromMe && contact.name === chatId && nameContact !== chatId) {
         this.logger.verbose('update contact name in chatwoot');
@@ -529,14 +506,20 @@ export class ChatwootService {
       if (contactConversations) {
         let conversation: any;
         if (this.provider.reopen_conversation) {
-          conversation = contactConversations.payload.find(
-            (conversation) => conversation.inbox_id == filterInbox.id,
-          );
+          conversation = contactConversations.payload.find((conversation) => conversation.inbox_id == filterInbox.id);
+
+          if (this.provider.conversation_pending) {
+            await client.conversations.toggleStatus({
+              accountId: this.provider.account_id,
+              conversationId: conversation.id,
+              data: {
+                status: 'pending',
+              },
+            });
+          }
         } else {
           conversation = contactConversations.payload.find(
-            (conversation) =>
-              conversation.status !== 'resolved' &&
-              conversation.inbox_id == filterInbox.id,
+            (conversation) => conversation.status !== 'resolved' && conversation.inbox_id == filterInbox.id,
           );
         }
         this.logger.verbose('return conversation if exists');
@@ -595,9 +578,7 @@ export class ChatwootService {
     }
 
     this.logger.verbose('find inbox by name');
-    const findByName = inbox.payload.find(
-      (inbox) => inbox.name === instance.instanceName,
-    );
+    const findByName = inbox.payload.find((inbox) => inbox.name === instance.instanceName);
 
     if (!findByName) {
       this.logger.warn('inbox not found');
@@ -699,8 +680,7 @@ export class ChatwootService {
 
     this.logger.verbose('find conversation by contact id');
     const conversation = findConversation.data.payload.find(
-      (conversation) =>
-        conversation?.meta?.sender?.id === contact.id && conversation.status === 'open',
+      (conversation) => conversation?.meta?.sender?.id === contact.id && conversation.status === 'open',
     );
 
     if (!conversation) {
@@ -820,8 +800,7 @@ export class ChatwootService {
 
     this.logger.verbose('find conversation by contact id');
     const conversation = findConversation.data.payload.find(
-      (conversation) =>
-        conversation?.meta?.sender?.id === contact.id && conversation.status === 'open',
+      (conversation) => conversation?.meta?.sender?.id === contact.id && conversation.status === 'open',
     );
 
     if (!conversation) {
@@ -871,12 +850,7 @@ export class ChatwootService {
     }
   }
 
-  public async sendAttachment(
-    waInstance: any,
-    number: string,
-    media: any,
-    caption?: string,
-  ) {
+  public async sendAttachment(waInstance: any, number: string, media: any, caption?: string) {
     this.logger.verbose('send attachment to instance: ' + waInstance.instanceName);
 
     try {
@@ -957,9 +931,7 @@ export class ChatwootService {
 
   public async receiveWebhook(instance: InstanceDto, body: any) {
     try {
-      this.logger.verbose(
-        'receive webhook to chatwoot instance: ' + instance.instanceName,
-      );
+      this.logger.verbose('receive webhook to chatwoot instance: ' + instance.instanceName);
       const client = await this.clientCw(instance);
 
       if (!client) {
@@ -972,8 +944,7 @@ export class ChatwootService {
 
       this.logger.verbose('check if is group');
       const chatId =
-        body.conversation.meta.sender?.phone_number?.replace('+', '') ||
-        body.conversation.meta.sender?.identifier;
+        body.conversation.meta.sender?.phone_number?.replace('+', '') || body.conversation.meta.sender?.identifier;
       const messageReceived = body.content;
       const senderName = body?.sender?.name;
       const waInstance = this.waMonitor.waInstances[instance.instanceName];
@@ -993,11 +964,7 @@ export class ChatwootService {
             await waInstance.connectToWhatsapp(number);
           } else {
             this.logger.verbose('whatsapp already connected');
-            await this.createBotMessage(
-              instance,
-              `🚨 ${body.inbox.name} instance is connected.`,
-              'incoming',
-            );
+            await this.createBotMessage(instance, `🚨 ${body.inbox.name} instance is connected.`, 'incoming');
           }
         }
 
@@ -1008,20 +975,12 @@ export class ChatwootService {
 
           if (!state) {
             this.logger.verbose('state not found');
-            await this.createBotMessage(
-              instance,
-              `⚠️ ${body.inbox.name} instance not found.`,
-              'incoming',
-            );
+            await this.createBotMessage(instance, `⚠️ ${body.inbox.name} instance not found.`, 'incoming');
           }
 
           if (state) {
             this.logger.verbose('state: ' + state + ' found');
-            await this.createBotMessage(
-              instance,
-              `⚠️ ${body.inbox.name} instance status: *${state}*`,
-              'incoming',
-            );
+            await this.createBotMessage(instance, `⚠️ ${body.inbox.name} instance status: *${state}*`, 'incoming');
           }
         }
 
@@ -1049,6 +1008,8 @@ export class ChatwootService {
             chatwoot_token: this.provider.token,
             chatwoot_url: this.provider.url,
             chatwoot_sign_msg: this.provider.sign_msg,
+            chatwoot_reopen_conversation: this.provider.reopen_conversation,
+            chatwoot_conversation_pending: this.provider.conversation_pending,
           };
 
           if (command.split(':')[2]) {
@@ -1070,19 +1031,10 @@ export class ChatwootService {
         }
       }
 
-      if (
-        body.message_type === 'outgoing' &&
-        body?.conversation?.messages?.length &&
-        chatId !== '123456'
-      ) {
+      if (body.message_type === 'outgoing' && body?.conversation?.messages?.length && chatId !== '123456') {
         this.logger.verbose('check if is group');
 
-        this.messageCacheFile = path.join(
-          ROOT_DIR,
-          'store',
-          'chatwoot',
-          `${instance.instanceName}_cache.txt`,
-        );
+        this.messageCacheFile = path.join(ROOT_DIR, 'store', 'chatwoot', `${instance.instanceName}_cache.txt`);
         this.logger.verbose('cache file path: ' + this.messageCacheFile);
 
         this.messageCache = this.loadMessageCache();
@@ -1103,9 +1055,7 @@ export class ChatwootService {
         if (senderName === null || senderName === undefined) {
           formatText = messageReceived;
         } else {
-          formatText = this.provider.sign_msg
-            ? `*${senderName}:*\n\n${messageReceived}`
-            : messageReceived;
+          formatText = this.provider.sign_msg ? `*${senderName}:*\n\n${messageReceived}` : messageReceived;
         }
 
         for (const message of body.conversation.messages) {
@@ -1119,12 +1069,7 @@ export class ChatwootService {
                 formatText = null;
               }
 
-              await this.sendAttachment(
-                waInstance,
-                chatId,
-                attachment.data_url,
-                formatText,
-              );
+              await this.sendAttachment(waInstance, chatId, attachment.data_url, formatText);
             }
           } else {
             this.logger.verbose('message is text');
@@ -1203,8 +1148,7 @@ export class ChatwootService {
       messageContextInfo: msg.messageContextInfo?.stanzaId,
       stickerMessage: undefined,
       documentMessage: msg.documentMessage?.caption,
-      documentWithCaptionMessage:
-        msg.documentWithCaptionMessage?.message?.documentMessage?.caption,
+      documentWithCaptionMessage: msg.documentWithCaptionMessage?.message?.documentMessage?.caption,
       audioMessage: msg.audioMessage?.caption,
       contactMessage: msg.contactMessage?.vcard,
       contactsArrayMessage: msg.contactsArrayMessage,
@@ -1405,24 +1349,14 @@ export class ChatwootService {
             }
 
             this.logger.verbose('send data to chatwoot');
-            const send = await this.sendData(
-              getConversion,
-              fileName,
-              messageType,
-              content,
-            );
+            const send = await this.sendData(getConversion, fileName, messageType, content);
 
             if (!send) {
               this.logger.warn('message not sent');
               return;
             }
 
-            this.messageCacheFile = path.join(
-              ROOT_DIR,
-              'store',
-              'chatwoot',
-              `${instance.instanceName}_cache.txt`,
-            );
+            this.messageCacheFile = path.join(ROOT_DIR, 'store', 'chatwoot', `${instance.instanceName}_cache.txt`);
 
             this.messageCache = this.loadMessageCache();
 
@@ -1436,24 +1370,14 @@ export class ChatwootService {
             this.logger.verbose('message is not group');
 
             this.logger.verbose('send data to chatwoot');
-            const send = await this.sendData(
-              getConversion,
-              fileName,
-              messageType,
-              bodyMessage,
-            );
+            const send = await this.sendData(getConversion, fileName, messageType, bodyMessage);
 
             if (!send) {
               this.logger.warn('message not sent');
               return;
             }
 
-            this.messageCacheFile = path.join(
-              ROOT_DIR,
-              'store',
-              'chatwoot',
-              `${instance.instanceName}_cache.txt`,
-            );
+            this.messageCacheFile = path.join(ROOT_DIR, 'store', 'chatwoot', `${instance.instanceName}_cache.txt`);
 
             this.messageCache = this.loadMessageCache();
 
@@ -1482,24 +1406,14 @@ export class ChatwootService {
           }
 
           this.logger.verbose('send data to chatwoot');
-          const send = await this.createMessage(
-            instance,
-            getConversion,
-            content,
-            messageType,
-          );
+          const send = await this.createMessage(instance, getConversion, content, messageType);
 
           if (!send) {
             this.logger.warn('message not sent');
             return;
           }
 
-          this.messageCacheFile = path.join(
-            ROOT_DIR,
-            'store',
-            'chatwoot',
-            `${instance.instanceName}_cache.txt`,
-          );
+          this.messageCacheFile = path.join(ROOT_DIR, 'store', 'chatwoot', `${instance.instanceName}_cache.txt`);
 
           this.messageCache = this.loadMessageCache();
 
@@ -1513,24 +1427,14 @@ export class ChatwootService {
           this.logger.verbose('message is not group');
 
           this.logger.verbose('send data to chatwoot');
-          const send = await this.createMessage(
-            instance,
-            getConversion,
-            bodyMessage,
-            messageType,
-          );
+          const send = await this.createMessage(instance, getConversion, bodyMessage, messageType);
 
           if (!send) {
             this.logger.warn('message not sent');
             return;
           }
 
-          this.messageCacheFile = path.join(
-            ROOT_DIR,
-            'store',
-            'chatwoot',
-            `${instance.instanceName}_cache.txt`,
-          );
+          this.messageCacheFile = path.join(ROOT_DIR, 'store', 'chatwoot', `${instance.instanceName}_cache.txt`);
 
           this.messageCache = this.loadMessageCache();
 
@@ -1580,16 +1484,9 @@ export class ChatwootService {
           return await this.createBotMessage(instance, erroQRcode, 'incoming');
         } else {
           this.logger.verbose('qrcode success');
-          const fileData = Buffer.from(
-            body?.qrcode.base64.replace('data:image/png;base64,', ''),
-            'base64',
-          );
+          const fileData = Buffer.from(body?.qrcode.base64.replace('data:image/png;base64,', ''), 'base64');
 
-          const fileName = `${path.join(
-            waInstance?.storePath,
-            'temp',
-            `${`${instance}.png`}`,
-          )}`;
+          const fileName = `${path.join(waInstance?.storePath, 'temp', `${`${instance}.png`}`)}`;
 
           this.logger.verbose('temp file name: ' + fileName);
 
@@ -1597,22 +1494,17 @@ export class ChatwootService {
           writeFileSync(fileName, fileData, 'utf8');
 
           this.logger.verbose('send qrcode to chatwoot');
-          await this.createBotQr(
-            instance,
-            'QRCode successfully generated!',
-            'incoming',
-            fileName,
-          );
+          await this.createBotQr(instance, 'QRCode successfully generated!', 'incoming', fileName);
 
           let msgQrCode = `⚡️ QRCode successfully generated!\n\nScan this QR code within the next 40 seconds.`;
 
           if (body?.qrcode?.pairingCode) {
             msgQrCode =
               msgQrCode +
-              `\n\n*Pairing Code:* ${body.qrcode.pairingCode.substring(
-                0,
+              `\n\n*Pairing Code:* ${body.qrcode.pairingCode.substring(0, 4)}-${body.qrcode.pairingCode.substring(
                 4,
-              )}-${body.qrcode.pairingCode.substring(4, 8)}`;
+                8,
+              )}`;
           }
 
           this.logger.verbose('send message to chatwoot');
@@ -1649,6 +1541,7 @@ export class ChatwootService {
         requestData['number'] = number;
       }
 
+      // eslint-disable-next-line
       const config = {
         method: 'post',
         maxBodyLength: Infinity,